refactor: reorganize sheduled service's package
