+++ src/test/java/com/example/spinlog/statistics/scheduled/startup/GenderStatisticsStartupServiceTest.java
@@ -1,4 +1,4 @@
-package com.example.spinlog.global.startup;
+package com.example.spinlog.statistics.scheduled.startup;
 
 import com.example.spinlog.article.entity.Emotion;
 import com.example.spinlog.article.entity.RegisterType;
@@ -11,7 +11,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.test.context.ActiveProfiles;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
